requirements.txt
selenium
beautifulsoup4
webdriver-manager
requests
python-dotenv
pandas
google-generativeai
